Intermediate
* Some missbehavior in Ability
* Costs are not fully parsed (Img.src)
Other parts looks function, but not fully testet

+++ src/main.java
@@ -1,17 +1,16 @@
 
 
+import model.CardRaw;
+
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 public class main {
     // private static DatabaseReference database;
 
     public static void main(String[] args) throws IOException{
-        boolean all = true;
+        boolean all = false;
 
         CardCsvWriter csv = null;
         CardJsonWriter json = null;
@@ -20,46 +19,49 @@ public class main {
         try {
             System.out.println("Starting:");
 
+            csv = null;
+            json = null; //new CardJsonWriter("newWalhalla.json");
+            fire = new CardFirebaseWriter("fowtest-f30af-service-account.json");
+
             File f = new File("newWalhalla.csv");
-            if (f.exists()) {
+            if (csv != null && f.exists()) {
                 f.delete();
             }
             f = new File("newWalhalla.json");
-            if (f.exists()) {
+            if (json != null && f.exists()) {
                 f.delete();
             }
-
-            csv = null;
-            json = null; // new CardJsonWriter("newWalhalla.json");
-            fire = new CardFirebaseWriter("fowtest-f30af-service-account.json");
 
             CardListRequest cardListRequest = new CardListRequest();
             List<CardRequest> cardRequests = null;
 
             do {
                 System.out.println(" Request page " + cardListRequest.page);
-                List<Card> cards = new CardRequester().request(
+                List<CardRaw> cardRaws = new CardRequester().request(
                         cardRequests = new CardListRequester().request(
                                 cardListRequest
                         )
                 );
 
-                System.out.println("  Got " + cards.size() + " cards:");
+                System.out.println("  Got " + cardRaws.size() + " cardRaws:");
 
-                for(int i = 0; i < cards.size(); i++) {
-                    System.out.println("   [" + i + "]: \"" + cards.get(i).name.de + "\" | \"" +
-                            cards.get(i).name.en + "\"");
-
+                for(int i = 0; i < cardRaws.size(); i++) {
                     if (csv != null) {
-                        csv.append(cards.get(i));
+                        csv.append(cardRaws.get(i));
                     }
                     if (json != null) {
-                        json.append(cards.get(i));
+                        json.append(cardRaws.get(i));
                     }
                     if(fire != null) {
-                        fire.append(cards.get(i));
+                        long time = fire.append(cardRaws.get(i));
+                        System.out.println("   [" + i + "]:{" + Long.toString(time) + "} \"" + cardRaws.get(i).name.de + "\" | \"" +
+                                cardRaws.get(i).name.en + "\"");
+                    } else {
+
+                        System.out.println("   [" + i + "]: \"" + cardRaws.get(i).name.de + "\" | \"" +
+                                cardRaws.get(i).name.en + "\"");
                     }
-                    // database.getRoot().push().setValue(cards.get(i), null);
+                    // database.getRoot().push().setValue(cardRaws.get(i), null);
                 }
 
                 cardListRequest.page++;
@@ -72,6 +74,11 @@ public class main {
             }
             if (json != null) {
                 json.close();
+            }
+            if (fire != null) {
+                try {
+                    fire.close();
+                } catch(Exception e){}
             }
         }
